requirements.txt
PyYAML
beautifulsoup4
openai
python-dotenv
selenium
tiktoken
webdriver-manager